+++ src/main/java/nu/aron/nextbuildnumber/Activation.java
@@ -1,0 +1,8 @@
+package nu.aron.nextbuildnumber;
+
+import org.apache.maven.execution.MavenSession;
+
+@FunctionalInterface
+interface Activation {
+    boolean test(MavenSession session, GetEnvPretender getEnvPretender);
+}