+++ src/main/java/nu/aron/nextbuildnumber/CommandInDirectory.java
@@ -8,8 +8,8 @@ import org.apache.maven.shared.utils.cli.Commandline;
 import java.io.File;
 
 interface CommandInDirectory {
-    static String run(File workingDirectory) {
-        Commandline cl = new Commandline("git rev-parse HEAD");
+    static String run(File workingDirectory, String command) {
+        Commandline cl = new Commandline(command);
         cl.setWorkingDirectory(workingDirectory);
         StringStreamConsumer stdout = new StringStreamConsumer();
         Try.of(() -> CommandLineUtils.executeCommandLine(cl, stdout, new LoggingConsumer()))