+++ src/main/java/nu/aron/nextbuildnumber/GetEnvPretender.java
@@ -1,0 +1,7 @@
+package nu.aron.nextbuildnumber;
+
+class GetEnvPretender {
+    String getenv(String key) {
+        return System.getenv(key);
+    }
+}