+++ src/main/java/nu/aron/nextbuildnumber/GitRevision.java
@@ -2,22 +2,20 @@ package nu.aron.nextbuildnumber;
 
 import org.apache.maven.execution.MavenSession;
 
-import java.io.File;
-
 import static nu.aron.nextbuildnumber.CommandInDirectory.run;
 import static nu.aron.nextbuildnumber.Constants.COMMIT;
 import static nu.aron.nextbuildnumber.Constants.log;
+import static nu.aron.nextbuildnumber.CurrentWorkingDirectory.getCwd;
 
 interface GitRevision {
 
     default void setRevision(MavenSession session) {
-        logAndSetProperty(session, run(new File(session.getRequest().getBaseDirectory())));
+        logAndSetProperty(session, run(getCwd(session), "git rev-parse HEAD"));
     }
 
     private void logAndSetProperty(MavenSession session, String value) {
         session.getSystemProperties().setProperty(COMMIT, value);
         session.getUserProperties().setProperty(COMMIT, value);
-//        session.getCurrentProject().getProperties().setProperty(COMMIT, value);
         log("System property {} set to {}", COMMIT, value);
     }
 }