+++ src/main/java/nu/aron/nextbuildnumber/LoggingConsumer.java
@@ -2,11 +2,11 @@ package nu.aron.nextbuildnumber;
 
 import org.apache.maven.shared.utils.cli.CommandLineUtils.StringStreamConsumer;
 
-import static nu.aron.nextbuildnumber.Constants.log;
+import static nu.aron.nextbuildnumber.Constants.logger;
 
 class LoggingConsumer extends StringStreamConsumer {
     @Override
     public void consumeLine(String line) {
-        log.error("{}", line);
+        logger.error("{}", line);
     }
 }