+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [10] - 2020-02-15
+### Added
+- Tries to be "CI build aware"
+- Takes no action if goals contains "deploy:deploy"
+- Adds branch name to version string if on a branch
+
 ## [8] - 2020-02-07
 ### Added
 - Now adds a property nextversion.commit to the deployed pom